# sdks/sdkwork-notary-app-sdk/openapi/notary-app-api.sdkgen.json
{
  "openapi": "3.1.2",
  "info": {
    "title": "SDKWork Notary App API",
    "version": "0.1.0",
    "description": "User-facing notary API. Merchandise owns sellable SPU/SKU matter definitions, Order owns checkout and order-item state, Payment remains behind Order for payment execution, Drive owns files and folders, and IAM owns organization members."
  },
  "servers": [
    {
      "url": "/app/v3/api"
    }
  ],
...
  "paths": {
    "/app/v3/api/notary/access": {
      "get": {
        "tags": [
          "notary"
        ],
        "operationId": "notary.access.retrieve",
        "summary": "Retrieve current notary business access",
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-owner": "sdkwork-notary",
        "x-sdkwork-api-authority": "sdkwork-notary.app",
        "x-sdkwork-domain": "notary",
        "x-sdkwork-resource": "notaryAccess",
        "x-sdkwork-permission": "notary.access.read",
        "x-sdkwork-tenant-scope": "tenant",
        "x-sdkwork-data-scope": "organization",
        "x-sdkwork-audit-event": "notary.access.read",
        "responses": {
          "200": {
            "description": "Current notary access",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SdkWorkApiResponse"
                    },
                    {
                      "type": "object",
                      "required": [
                        "data"
                      ],
                      "properties": {
                        "data": {
                          "type": "object",
                          "required": [
                            "item"
                          ],
                          "properties": {
                            "item": {
                              "$ref": "#/components/schemas/NotaryAccess"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Problem detail",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Problem detail",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "app-api",
        "x-sdkwork-auth-mode": "dual-token",
        "x-sdkwork-source-route-crate": "sdkwork-routes-notary-app-api",
        "x-sdkwork-source": "crates/sdkwork-routes-notary-app-api/src/handlers.rs"
      }
    },
...
    "/app/v3/api/notary/cases/{caseId}": {
      "get": {
        "tags": [
          "notary"
        ],
        "operationId": "notary.cases.retrieve",
        "summary": "Retrieve a notary case",
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-owner": "sdkwork-notary",
        "x-sdkwork-api-authority": "sdkwork-notary.app",
        "x-sdkwork-domain": "notary",
        "x-sdkwork-resource": "notaryCase",
        "x-sdkwork-permission": "notary.cases.read",
        "x-sdkwork-tenant-scope": "tenant",
        "x-sdkwork-data-scope": "organization",
        "x-sdkwork-audit-event": "notary.cases.retrieve",
        "parameters": [
          {
            "$ref": "#/components/parameters/CaseId"
          }
        ],
        "responses": {
          "200": {
            "description": "Notary case",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SdkWorkApiResponse"
                    },
                    {
                      "type": "object",
                      "required": [
                        "data"
                      ],
                      "properties": {
                        "data": {
                          "type": "object",
                          "required": [
                            "item"
                          ],
                          "properties": {
                            "item": {
                              "$ref": "#/components/schemas/NotaryCase"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Problem detail",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Problem detail",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Problem detail",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "app-api",
        "x-sdkwork-auth-mode": "dual-token",
        "x-sdkwork-source-route-crate": "sdkwork-routes-notary-app-api",
        "x-sdkwork-source": "crates/sdkwork-routes-notary-app-api/src/handlers.rs"
      },
      "patch": {
        "tags": [
          "notary"
        ],
        "operationId": "notary.cases.update",
        "summary": "Update mutable notary case fields",
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-owner": "sdkwork-notary",
        "x-sdkwork-api-authority": "sdkwork-notary.app",
        "x-sdkwork-domain": "notary",
        "x-sdkwork-resource": "notaryCase",
        "x-sdkwork-permission": "notary.cases.update",
        "x-sdkwork-tenant-scope": "tenant",
        "x-sdkwork-data-scope": "organization",
        "x-sdkwork-audit-event": "notary.cases.update",
        "parameters": [
          {
            "$ref": "#/components/parameters/CaseId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNotaryCaseRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Notary case",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SdkWorkApiResponse"
                    },
                    {
                      "type": "object",
                      "required": [
                        "data"
                      ],
                      "properties": {
                        "data": {
                          "type": "object",
                          "required": [
                            "item"
                          ],
                          "properties": {
                            "item": {
                              "$ref": "#/components/schemas/NotaryCase"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "400": {
            "description": "Problem detail",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Problem detail",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Problem detail",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "404": {
            "description": "Problem detail",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "409": {
            "description": "Problem detail",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "app-api",
        "x-sdkwork-auth-mode": "dual-token",
        "x-sdkwork-source-route-crate": "sdkwork-routes-notary-app-api",
        "x-sdkwork-source": "crates/sdkwork-routes-notary-app-api/src/handlers.rs"
      }
    },
...
    "/app/v3/api/notary/dashboard/statistics": {
      "get": {
        "tags": [
          "notary"
        ],
        "operationId": "notary.dashboard.statistics.retrieve",
        "summary": "Retrieve notary dashboard statistics",
        "description": "Returns pending review queue count, today completed count, anomaly intercepted count, and monthly preservation total for the current organization",
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-owner": "sdkwork-notary",
        "x-sdkwork-api-authority": "sdkwork-notary.app",
        "x-sdkwork-domain": "notary",
        "x-sdkwork-resource": "notaryDashboardStatistics",
        "x-sdkwork-permission": "notary.dashboard.read",
        "x-sdkwork-tenant-scope": "tenant",
        "x-sdkwork-data-scope": "organization",
        "x-sdkwork-audit-event": "notary.dashboard.statistics.retrieve",
        "responses": {
          "200": {
            "description": "Notary dashboard statistics",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SdkWorkApiResponse"
                    },
                    {
                      "type": "object",
                      "required": [
                        "data"
                      ],
                      "properties": {
                        "data": {
                          "type": "object",
                          "required": [
                            "item"
                          ],
                          "properties": {
                            "item": {
                              "$ref": "#/components/schemas/NotaryStatistics"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Problem detail",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Problem detail",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "app-api",
        "x-sdkwork-auth-mode": "dual-token",
        "x-sdkwork-source-route-crate": "sdkwork-routes-notary-app-api",
        "x-sdkwork-source": "crates/sdkwork-routes-notary-app-api/src/handlers.rs"
      }
    },
    "/app/v3/api/notary/reports/monthly": {
      "get": {
        "tags": [
          "notary"
        ],
        "operationId": "notary.reports.monthly.retrieve",
        "summary": "Download monthly business report",
        "description": "Generate and download monthly business report for notary operations",
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-owner": "sdkwork-notary",
        "x-sdkwork-api-authority": "sdkwork-notary.app",
        "x-sdkwork-domain": "notary",
        "x-sdkwork-resource": "notaryMonthlyReport",
        "x-sdkwork-permission": "notary.reports.read",
        "x-sdkwork-tenant-scope": "tenant",
        "x-sdkwork-data-scope": "organization",
        "x-sdkwork-audit-event": "notary.reports.monthly.retrieve",
        "parameters": [
          {
            "name": "month",
            "in": "query",
            "required": false,
            "description": "Month in YYYY-MM format, defaults to current month",
            "schema": {
              "type": "string",
              "pattern": "^\\d{4}-\\d{2}$",
              "example": "2026-06"
            }
          },
          {
            "name": "format",
            "in": "query",
            "required": false,
            "description": "Report format",
            "schema": {
              "type": "string",
              "enum": [
                "pdf",
                "excel",
                "csv"
              ],
              "default": "pdf"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Monthly business report download",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SdkWorkApiResponse"
                    },
                    {
                      "type": "object",
                      "required": [
                        "data"
                      ],
                      "properties": {
                        "data": {
                          "type": "object",
                          "required": [
                            "item"
                          ],
                          "properties": {
                            "item": {
                              "$ref": "#/components/schemas/MonthlyReport"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "401": {
            "description": "Problem detail",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "403": {
            "description": "Problem detail",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-api-surface": "app-api",
        "x-sdkwork-auth-mode": "dual-token",
        "x-sdkwork-source-route-crate": "sdkwork-routes-notary-app-api",
        "x-sdkwork-source": "crates/sdkwork-routes-notary-app-api/src/handlers.rs"
      }
    }
  },
  "components": {
    "securitySchemes": {
      "AuthToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "SDKWork auth token carried as Authorization: Bearer <auth_token>."
      },
      "AccessToken": {
        "type": "apiKey",
        "in": "header",
        "name": "Access-Token",
        "description": "SDKWork access isolation token."
      }
    },
    "parameters": {
...
      "CaseId": {
        "name": "caseId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
...
    },
...
    "schemas": {
      "ProblemDetail": {
        "type": "object",
        "additionalProperties": true,
        "required": [
          "type",
          "title",
          "status",
          "code",
          "traceId"
        ],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri-reference"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "minimum": 100,
            "maximum": 599
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "code": {
            "$ref": "#/components/schemas/SdkWorkPlatformErrorCode"
          },
          "traceId": {
            "type": "string",
            "format": "uuid",
            "description": "Server-owned request correlation id."
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          }
        }
      },
...
      "NotaryAccess": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "visible",
          "organizationVerified",
          "notaryBusinessEnabled",
          "memberId",
          "permissions"
        ],
        "properties": {
          "visible": {
            "type": "boolean"
          },
          "organizationVerified": {
            "type": "boolean"
          },
          "notaryBusinessEnabled": {
            "type": "boolean"
          },
          "tenantId": {
            "type": "string"
          },
          "organizationId": {
            "type": "string"
          },
          "memberId": {
            "type": "string"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "positions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "departments": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "reason": {
            "type": "string"
          }
        }
      },
...
      "NotaryCase": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "caseNo",
          "createTime",
          "applicant",
          "title",
          "notary",
          "remarks",
          "type",
          "status",
          "fee",
          "orderId",
          "orderItemId",
          "skuId",
          "driveSpaceId",
          "driveSpaceType",
          "driveFolderNodeId",
          "parties",
          "documents",
          "timeline"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "caseNo": {
            "type": "string"
          },
          "createTime": {
            "type": "string",
            "format": "date-time"
          },
          "processTime": {
            "type": "string",
            "format": "date-time"
          },
          "applicant": {
            "type": "string"
          },
          "applicantUserId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "notary": {
            "type": "string"
          },
          "primaryNotaryMembershipId": {
            "type": "string"
          },
          "remarks": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/NotaryCaseStatus"
          },
          "fee": {
            "type": "string",
            "x-sdkwork-decimal-string": true
          },
          "currencyCode": {
            "type": "string"
          },
          "hash": {
            "type": "string"
          },
          "orderId": {
            "type": "string"
          },
          "orderItemId": {
            "type": "string"
          },
          "skuId": {
            "type": "string"
          },
          "driveSpaceId": {
            "type": "string"
          },
          "driveSpaceType": {
            "type": "string",
            "enum": [
              "notary"
            ]
          },
          "driveFolderNodeId": {
            "type": "string"
          },
          "parties": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Party"
            }
          },
          "documents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NotaryDocument"
            }
          },
          "timeline": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TimelineEvent"
            }
          },
          "version": {
            "type": "string",
            "format": "int64",
            "x-sdkwork-int64-string": true
          }
        }
      },
...
      "UpdateNotaryCaseRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200
          },
          "description": {
            "type": "string",
            "maxLength": 2000
          },
          "remarks": {
            "type": "string",
            "maxLength": 2000
          },
          "version": {
            "type": "string",
            "format": "int64",
            "x-sdkwork-int64-string": true
          }
        }
      },
...
      "NotaryStatistics": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "pendingReviewQueue",
          "todayCompleted",
          "anomalyIntercepted",
          "monthlyPreservationTotal"
        ],
        "properties": {
          "pendingReviewQueue": {
            "type": "object",
            "required": [
              "count"
            ],
            "properties": {
              "count": {
                "type": "integer",
                "description": "Number of cases pending review"
              },
              "estimatedProcessHours": {
                "type": "number",
                "format": "float",
                "description": "Estimated hours to process pending cases"
              }
            }
          },
          "todayCompleted": {
            "type": "object",
            "required": [
              "count"
            ],
            "properties": {
              "count": {
                "type": "integer",
                "description": "Number of cases completed today"
              },
              "comparedToYesterday": {
                "type": "integer",
                "description": "Difference compared to yesterday"
              }
            }
          },
          "anomalyIntercepted": {
            "type": "object",
            "required": [
              "count"
            ],
            "properties": {
              "count": {
                "type": "integer",
                "description": "Number of anomaly cases intercepted by risk control"
              },
              "interceptorType": {
                "type": "string",
                "description": "Type of risk control interceptor"
              }
            }
          },
          "monthlyPreservationTotal": {
            "type": "object",
            "required": [
              "count"
            ],
            "properties": {
              "count": {
                "type": "integer",
                "description": "Total number of cases preserved this month"
              },
              "blockchainSyncStatus": {
                "type": "string",
                "enum": [
                  "OK",
                  "PENDING",
                  "ERROR"
                ],
                "default": "OK",
                "description": "Blockchain synchronization status"
              }
            }
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "description": "When the statistics were computed"
          }
        }
      },
      "MonthlyReport": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "downloadUrl",
          "reportId"
        ],
        "properties": {
          "downloadUrl": {
            "type": "string",
            "format": "uri",
            "description": "URL to download the report file"
          },
          "reportId": {
            "type": "string",
            "description": "Unique report identifier"
          },
          "month": {
            "type": "string",
            "description": "Month the report covers in YYYY-MM format"
          },
          "format": {
            "type": "string",
            "enum": [
              "pdf",
              "excel",
              "csv"
            ],
            "description": "Report file format"
          },
          "generatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "When the report was generated"
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time",
            "description": "When the download URL expires"
          },
          "fileSize": {
            "type": "integer",
            "description": "Report file size in bytes"
          },
          "caseCount": {
            "type": "integer",
            "description": "Number of cases included in the report"
          }
        }
      },
      "SdkWorkApiResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "data",
          "traceId"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "enum": [
              0
            ],
            "default": 0,
            "minimum": 0,
            "maximum": 0
          },
          "data": {
            "description": "Operation-specific payload typed per response schema."
          },
          "traceId": {
            "type": "string",
            "format": "uuid",
            "description": "Server-owned request correlation id."
          }
        }
      },
...
      "SdkWorkPlatformErrorCode": {
        "type": "integer",
        "format": "int32",
        "minimum": 40001,
        "maximum": 79999,
        "description": "Platform or domain error code per API_SPEC.md §15.3."
      },
...
    }
  }
}